[250] Silvermoon City - Event in "The Bazzar"
WoWWiki:
(...)Lyria Skystrider is a blood elf found in The Bazaar of Silvermoon
City. She, along with Priest Ennas, is leading a public forum arguing
against the policies of the blood elven leadership, specifically their
decision to break off their ties to the Alliance and join the Horde. She
proclaims that Sylvanas Windrunner is not the same one that they once
knew, the orcs are "fair-weather friends", and the magisters are
traitors. (...)