[1450][ACID] Warsong Hold Invasion EVENT
- Warsong Hold Invasion EVENT
- Repeats every 30mins.

NOTE:
Theres confusion what triggers this event to start. According to wowhead
.. it may start with new 68-69 player comming first time to Warsong
Hold. Other is static 30min timer ...that from time to time - may be
canceled by smth...
CMaNGoS - is created for not massive population ..so second option is
better.. and this is how Script was made:
Repeats every 30 mins + 2 conditions must be met:
- only One Invis trigger can be present at same time
- final Boss from prev invasion must be killed

Part of:
https://github.com/cmangos/wotlk-db/issues/231

Closes:
https://github.com/cmangos/wotlk-db/issues/210