[INS] Instance File System Implemented
Instance files are a new method of tracking all the data we have which
pertains to the creatures/objects within all dungeons and raids.
Each file corresponds to an individual instance and the first three
digits of the filename is the official number of that map.

Thx to TBCDB team for hard work with this...
Source:

https://github.com/cmangos/tbc-db/commit/1f2cfc607b48030eea9902dd3ceee65cf193a8c6

*NOTE* this is only BASE .. individual instance files will be added
later

Also:
- InstallFullDB.sh file updated
- Unused files removed  - thx to @Phatcat

Closes:
https://github.com/cmangos/wotlk-db/pull/288